refactor : 거리 정책 패키지 생성 & 이동
