+++ src/main/java/nextstep/subway/path/domain/policy/distance/OverTenDistancePolicy.java
@@ -1,7 +1,9 @@
-package nextstep.subway.path.domain.policy;
+package nextstep.subway.path.domain.policy.distance;
+
+import nextstep.subway.path.domain.policy.FarePolicy;
 
 import static nextstep.subway.path.domain.Fare.BASIC_FARE;
-import static nextstep.subway.path.domain.policy.DistancePolicyFactory.TEN_DISTANCE;
+import static nextstep.subway.path.domain.policy.distance.DistancePolicyFactory.TEN_DISTANCE;
 import static nextstep.subway.path.domain.policy.FarePolicy.calculateOverFare;
 
 public class OverTenDistancePolicy implements FarePolicy {
