+++ src/test/java/nextstep/subway/path/domain/policy/DefaultDistancePolicyTest.java
@@ -1,10 +1,9 @@
 package nextstep.subway.path.domain.policy;
 
+import nextstep.subway.path.domain.policy.distance.DefaultDistancePolicy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class DefaultDistancePolicyTest {
