+++ src/test/java/nextstep/subway/path/domain/policy/OverTenDistancePolicyTest.java
@@ -1,5 +1,6 @@
 package nextstep.subway.path.domain.policy;
 
+import nextstep.subway.path.domain.policy.distance.OverTenDistancePolicy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
